requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-login==0.6.2
pymysql==1.1.0
requests==2.31.0
python-dotenv==1.0.0
werkzeug==2.3.7
pyjwt==2.8.0

